Rename left panel slice
