+++ frontend/src/context/store.ts
@@ -15,7 +15,7 @@ import mapSelectionLayerStateReduce from './mapSelectionLayerStateSlice';
 import mapBoundaryInfoStateReduce from './mapBoundaryInfoStateSlice';
 import datasetResultStateReduce from './datasetStateSlice';
 import mapTileLoadingStateReduce from './mapTileLoadingStateSlice';
-import sidebarStateReduce from './sidebarStateSlice';
+import leftPanelStateReduce from './leftPanelStateSlice';
 
 const reducer = combineReducers({
   mapState: mapStateReduce,
@@ -28,7 +28,7 @@ const reducer = combineReducers({
   mapBoundaryInfoState: mapBoundaryInfoStateReduce,
   datasetState: datasetResultStateReduce,
   mapTileLoadingState: mapTileLoadingStateReduce,
-  sidebarState: sidebarStateReduce,
+  leftPanelState: leftPanelStateReduce,
 });
 
 export const store = configureStore({
